chore: relocate Select component
The Select component was modularized and relocated.
