+++ src/components/select/index.ts
@@ -1,0 +1,1 @@
+export { default as Select } from "./Select";
